refactor: move context to separate file
