+++ src/context/themeProvider.tsx
@@ -1,0 +1,23 @@
+import { ReactNode, useState } from 'react';
+import { ThemeContext } from './themeContext';
+
+type Theme = 'light' | 'dark';
+
+interface ThemeProviderProps {
+  children: ReactNode;
+}
+
+export const ThemeProvider = ({ children }: ThemeProviderProps) => {
+  const [theme, setTheme] = useState<Theme>('dark');
+
+  const toggleTheme = () => {
+    if (theme == 'light') setTheme('dark');
+    else setTheme('light');
+  };
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
